Send tps on cooldown start, fix overlay being applied multiple times per frame

+++ src/main/java/org/darkoro/zerosmod/zsweapons/network/CooldownToClientPacket.java
@@ -6,10 +6,10 @@ import io.netty.buffer.ByteBuf;
 public class CooldownToClientPacket implements IMessage {
     public int fullCooldown;
     public int cooldown;
-    public float tps;
+    public double tps;
 
     public CooldownToClientPacket() {}
-    public CooldownToClientPacket(int fullCooldown, int cooldown, float tps) {
+    public CooldownToClientPacket(int fullCooldown, int cooldown, double tps) {
         this.fullCooldown = fullCooldown;
         this.cooldown = cooldown;
         this.tps = tps;
@@ -18,12 +18,12 @@ public class CooldownToClientPacket implements IMessage {
     @Override public void fromBytes(ByteBuf buf) {
         fullCooldown = buf.readInt();
         cooldown = buf.readInt();
-        tps = buf.readFloat();
+        tps = buf.readDouble();
     }
 
     @Override public void toBytes(ByteBuf buf) {
         buf.writeInt(fullCooldown);
         buf.writeInt(cooldown);
-        buf.writeFloat(tps);
+        buf.writeDouble(tps);
     }
 }
